# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Replit Multi-Account Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/accounts:
    get:
      operationId: getProjectAccounts
      tags: [projects]
      summary: Get accounts linked to project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Linked accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: linkAccountToProject
      tags: [projects]
      summary: Link an account to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectAccountLink"
      responses:
        "200":
          description: Linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /projects/{id}/accounts/{accountId}:
    delete:
      operationId: unlinkAccountFromProject
      tags: [projects]
      summary: Unlink an account from a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unlinked
...
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List skills library
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Create a skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{id}:
    patch:
      operationId: updateSkill
      tags: [skills]
      summary: Update skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        tokenMasked:
          type: string
        status:
          type: string
          enum: [active, inactive, error]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - username
        - tokenMasked
        - status
        - createdAt
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        replUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paused, completed]
        accountCount:
          type: integer
        memoryCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - status
        - accountCount
        - memoryCount
        - createdAt
        - updatedAt
    ProjectInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        replUrl:
          type: string
      required:
        - name
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        replUrl:
          type: string
        status:
          type: string
    ProjectAccountLink:
      type: object
      properties:
        accountId:
          type: integer
      required:
        - accountId
...
    Skill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        prompt:
          type: string
        category:
          type: string
        usageCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - prompt
        - category
        - usageCount
        - createdAt
    SkillInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        prompt:
          type: string
        category:
          type: string
      required:
        - name
        - description
        - prompt
        - category
    SkillUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        prompt:
          type: string
        category:
          type: string
...